Improved robustness of optimization with two features:   - Remove small disconnected volume components (can happen in some rare, weird     merged cases).   - Randomly perturb parameters on iterate construction failure and retry a few     times before failing. Tool to replicate orthotropic base cell and its fields into an "Nx x Ny x Nz" periodic tiling (where sizes are powers of two).

+++ isosurface_inflator/PrintabilityCheck.cc
@@ -90,7 +90,7 @@ int main(int argc, const char *argv[]) {
 
         Eigen::VectorXd cval = C * paramVec;
         bool satisfied = true;
-        for (size_t crow = 0; crow < C.rows(); ++crow) {
+        for (size_t crow = 0; crow < size_t(C.rows()); ++crow) {
             if (cval[crow] < -1e-12) satisfied = false;
         }
         if (satisfied != isPrintable[i]) {
